Add LXA TAC external electrical test equipment support
The LXA TAC external electrical test (eet) equipment is a specialized
device that is capable of multiplexing signal from and to the TAC, some
shunt resistors and external power supplies.

It is connected to the exporter via USB and gets represented as an I2C
device. The lxatac-eet resources, driver and agent are used to control
the device from the system running the tests.

To select tests using the eet only when it is present, the feature "eet"
is introduced.

Since upstream labgrid does not know of this device, the resource's
information are not provided by the exporter. Thus the
`LxatacEETResource` must describe host and USB path of the resource.