test/userspace: Mark PTX-internal tests with a labgrid feature
PTX works are volumes on which developers at Pengutronix do their actual
work. These volumes are mounted on all labgrid exporters via NFS.
This way files are directly available on the exporter - without the need
of an copying step.

This feature is only configured in our internal PTX-flavor of
`meta-lxatac`. So let's mark this test as only relevant for the
PTX-flavor.