strategy: integrate ethmux driver
In order to add the Ethmux (that provides an `DigitalOutputProtocol`),
we need to name the other `DigitalOutputProtocol`-resource as well.
Thus this commit adds names and named bindings for the `dfu_mode`
resource as well.

When the environment defines an Ethmux we want to initialize it in the
strategy before booting. This way we make sure that the DUT has network
connectivity, regardless of the previous state of the Ethmux.