lxatac-eet: Use full path for i2c-tools
The debian package `i2c-tools` package provides `i2cset` in `/usr/sbin`
and thus a normal user will probably not have it in their ´PATH`.

But with the udev-rule given in `contrib/99-tiny-i2c.rules` we will give
`plugdev` access to the pluggable I2C-devices and we can simply use
`i2cset` - at least on this one `i2c-dev`.