env/lxatac-vanilla-eet: Use FQDN for network service
The runner executing the test suite are not part of
`lab2.stw.pengutronix.de` and thus do not have `lab2` in their search
domain.
Using the FQDN the runners can reach the DUT in `lab2` as intended.

This is actually how it is done in `lxatac-ptx.yml` since it's first
commit.
So this env probably only always worked on the development machines
(which have `lab2` in their search domains) but never on the runners.