test_linux_hardware: Test for I2C devices instead of the subsystem
The previous test very bluntly only tested if the I2C-Subsystem in the
kernel has been loaded at all. But this does not tell us anything about
the actual devices on the bus we want to use.

The new tests instead try to use the I2C devices the way userspace
would.
Actually none of the I2C devices are directly used by userspace:

* The EEPROMs are read by the bootloader and their information
  are added to the devicetree's chosen_node.
  But for the EEPROMs we can at least do a read, so we can check
  if this works.
* The PMIC is used inside the kernel.
  So we will only check the hardware has been probed.
* The USB-Hub is also used inside the kernel
  So we will only check the hardware has been probed.