test_interface_usb: Disconnect USB drive to eliminate buffering
By disconnecting the USB drive during the test we can be sure that Linux
has dropped all buffers for the USB storage.
And thus that this test not only tests buffers somewhere in the kernel
without actually writing data to the USB drive.

Since we already have the _eet_ connected to the TAC we can simply use
it to disconnect and connect the USB-stick during the test.