test_interface_usb_io: Do not infer block device from USB path
Inferring the block device from the USB path at runtime is not needed,
since the path in `/dev/disk/by-path/` will be stable for a given USB
storage device in a given USB port.

This way we get rid of a few extra commands.
But more important: We get rid of pipes in shell commands, that may fail
without failing the whole command.