test_network/prepare_network: Make sure we are online at the end of test
The fixture `prepare_network` tears down the whole network setup of the
TAC and replaces it with it's own setup. At the end of the test the
process is reversed.

But to make sure that the TAC is actually online we should at least wait
until we have reached the same conditions as in the `network` state of
the strategy.
This way, if following tests rely on the TAC being online we can still
fulfill that.