test_filesystems: Only test in base system
Until now we have tested the size of the file systems in `slot 0` and
`slot 1`. This seems natural, since these are two independent systems.

But both slots basically contain the same system - only installed via
two different methods.
We currently have no reason to believe that the system installed via
RAUC may have another file system layout that the system installed via
fastboot.

Thus we rewrite the tests to only test the file system sizes in `slot 0`
and remove a good amount of complexity and testing time.
If we ever encounter problems in this realm we can still write tests
that focus on the actual problems fixed.