Check for failed units in strategy instead of tests
The strategy already checks, whether the system is running.
Thus, it makes total sense to also collect debug information if this
state could not be reached.

This is a preparation for the removal of RAUC handling from the
strategy.
After support for RAUC  has been removed from the strategy, the strategy
will always assume, that we are in `system0`.
Thus it does not make sense to have tests for `system1` anymore.