Move RAUC handling from strategy into tests
Until now we handling of RAUC installs and handling of state transitions
between these two systems was done by the strategy.
This led to a fair amount of added complexity in the strategy.
For example did some states pretend to be reached (shell, network), even
if actually a we were in one of the RAUC-states (`system0`, `system1`).

Reducing the complexity in the strategy makes it also clear that other
special cases, like reconfiguring network from `online` to `loopback`
should not be part of the strategy.

This commit removes handling of RAUC installation and tracking of the
active slot from the strategy.
As a replacement it adds pytest `fixtures` that allow test cases
to manipulate which slot should be booted.

The fixture `set_bootstate_in_bootloader` should be used in tests that
want to boot into `system1` or change retries, `mark-good` etc.
This fixture makes sure (using the `default_bootstate` fixture) that the
system will always be in `system0` and has the default bootste after the
test has finished.

The only requirements for tests is, that they should not install a new
system into `system0`, nor make `system0` unusable.

While doing all these changes the actual RAUC tests have been adapted to
more closely follow the best current practices for these tests.