tests: Use eet via fixture to make connections will be reverted
Without this change all tests left the *eet* in what ever state happend
to be the last. This does not work well with the notion that tests
should leave the system in it's initial state.

With this change the new `eet`-fixture will make sure that the *eet* is
returned to "no connection" at tear down.