strategy: use a longer login timeout on the first boot
The first boot is special because it is in fact two boots before we can
log in: the first one re-partitions the eMMC and sets up the data
partition, then the TAC reboots and the second boot is the one we can
actually log in to.

Sometimes this finishes in the default login time of 60 seconds,
sometimes it does not.

Make the test more stable by using a 5 min login timeout just for this
first boot.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>