test_network/tftp: Log duration of get and put actions
This test has been flaky in the past, due to timeouts during get and
put.
But it's unclear if this is due to changes in the system, or if the
timeout was a close call in the first place.
Let's add this as a measurement, so we can take a look at this value
over time, if needed.