tests/can: Make can_configured more universal
This commit moves the link up from the fixture to the tests.
This way this fixture can also be used in tests that only want one
interface up.

We also make sure that both interfaces are set down during preparation
and tear-down. Doing so, we can reset the berr-counters of both
interfaces.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>