tests/can: Test if resetting berr-counters works
Resetting the berr-counters in the CAN peripherals on STM32MP1 is only
possible when both interfaces are down - since both peripherals share
the same reset line.

With this test we make sure that we can at least reset one interface as
long as the other is already down.

This change needs
https://github.com/linux-automation/meta-lxatac/pull/284
to be running on the DUT.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>