test/test_tacd_dut_power_switchable: Increase analog waiting time
This aligns with other increased delays like e.g. the one in commit
a518273 ("test/test_tacd_dut_power_off_floating: Increase analog waiting
time").

This was the only place where we still had 0.1s of delay and this is a
location where we sometimes see issues with failed assertions.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>